- Split reactive and sync contexts - Support 2.4.1 features
