+++ spring-boot-autoconfigure-data-aerospike/src/main/java/org/springframework/boot/autoconfigure/data/aerospike/AerospikeDataConfiguration.java
@@ -17,32 +17,21 @@
 package org.springframework.boot.autoconfigure.data.aerospike;
 
 import com.aerospike.client.AerospikeClient;
-import org.springframework.beans.BeanUtils;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
-import org.springframework.boot.autoconfigure.domain.EntityScanner;
-import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.data.aerospike.convert.AerospikeCustomConversions;
-import org.springframework.data.aerospike.convert.AerospikeTypeAliasAccessor;
+import org.springframework.context.annotation.Lazy;
 import org.springframework.data.aerospike.convert.MappingAerospikeConverter;
 import org.springframework.data.aerospike.core.AerospikeExceptionTranslator;
 import org.springframework.data.aerospike.core.AerospikeTemplate;
-import org.springframework.data.aerospike.core.DefaultAerospikeExceptionTranslator;
+import org.springframework.data.aerospike.index.AerospikePersistenceEntityIndexCreator;
 import org.springframework.data.aerospike.mapping.AerospikeMappingContext;
-import org.springframework.data.aerospike.mapping.Document;
 import org.springframework.data.aerospike.query.QueryEngine;
 import org.springframework.data.aerospike.query.StatementBuilder;
 import org.springframework.data.aerospike.query.cache.IndexInfoParser;
 import org.springframework.data.aerospike.query.cache.IndexRefresher;
-import org.springframework.data.aerospike.query.cache.IndexesCache;
-import org.springframework.data.aerospike.query.cache.IndexesCacheHolder;
 import org.springframework.data.aerospike.query.cache.IndexesCacheUpdater;
 import org.springframework.data.aerospike.query.cache.InternalIndexOperations;
-import org.springframework.data.annotation.Persistent;
-import org.springframework.data.mapping.model.FieldNamingStrategy;
-
-import java.util.Collections;
 
 /**
  * Configure Spring Data's Aerospike support.
@@ -51,7 +40,7 @@ import java.util.Collections;
  * @author Anastasiia Smirnova
  */
 @Configuration(proxyBeanMethods = false)
-public class AerospikeDataConfiguration {
+class AerospikeDataConfiguration {
 
     @Bean(name = "aerospikeTemplate")
     @ConditionalOnMissingBean(name = "aerospikeTemplate")
@@ -78,18 +67,6 @@ public class AerospikeDataConfiguration {
         return queryEngine;
     }
 
-    @Bean(name = "aerospikeStatementBuilder")
-    @ConditionalOnMissingBean(name = "aerospikeStatementBuilder")
-    public StatementBuilder aerospikeStatementBuilder(IndexesCache indexesCache) {
-        return new StatementBuilder(indexesCache);
-    }
-
-    @Bean(name = "aerospikeIndexCache")
-    @ConditionalOnMissingBean(name = "aerospikeIndexCache")
-    public IndexesCacheHolder aerospikeIndexCache() {
-        return new IndexesCacheHolder();
-    }
-
     @Bean(name = "aerospikeIndexRefresher")
     @ConditionalOnMissingBean(name = "aerospikeIndexRefresher")
     public IndexRefresher aerospikeIndexRefresher(AerospikeClient aerospikeClient, IndexesCacheUpdater indexesCacheUpdater) {
@@ -98,46 +75,10 @@ public class AerospikeDataConfiguration {
         return refresher;
     }
 
-    @Bean(name = "mappingAerospikeConverter")
-    @ConditionalOnMissingBean(name = "mappingAerospikeConverter")
-    public MappingAerospikeConverter mappingAerospikeConverter(AerospikeMappingContext aerospikeMappingContext,
-                                                               AerospikeTypeAliasAccessor aerospikeTypeAliasAccessor,
-                                                               AerospikeCustomConversions aerospikeCustomConversions) {
-        return new MappingAerospikeConverter(aerospikeMappingContext, aerospikeCustomConversions, aerospikeTypeAliasAccessor);
+    @Bean
+    @ConditionalOnMissingBean(name = "aerospikePersistenceEntityIndexCreator")
+    public AerospikePersistenceEntityIndexCreator aerospikePersistenceEntityIndexCreator(AerospikeMappingContext aerospikeMappingContext,
+                                                                                         @Lazy AerospikeTemplate template) {
+        return new AerospikePersistenceEntityIndexCreator(aerospikeMappingContext, template);
     }
-
-    @Bean(name = "aerospikeTypeAliasAccessor")
-    @ConditionalOnMissingBean(name = "aerospikeTypeAliasAccessor")
-    public AerospikeTypeAliasAccessor aerospikeTypeAliasAccessor(AerospikeDataProperties aerospikeDataProperties) {
-        return new AerospikeTypeAliasAccessor(aerospikeDataProperties.getTypeKey());
-    }
-
-    @Bean(name = "aerospikeCustomConversions")
-    @ConditionalOnMissingBean(name = "aerospikeCustomConversions")
-    public AerospikeCustomConversions aerospikeCustomConversions() {
-        return new AerospikeCustomConversions(Collections.emptyList());
-    }
-
-    @Bean(name = "aerospikeMappingContext")
-    @ConditionalOnMissingBean(name = "aerospikeMappingContext")
-    public AerospikeMappingContext aerospikeMappingContext(ApplicationContext applicationContext,
-                                                           AerospikeCustomConversions aerospikeCustomConversions,
-                                                           AerospikeDataProperties aerospikeDataProperties) throws Exception {
-        AerospikeMappingContext context = new AerospikeMappingContext();
-        context.setInitialEntitySet(new EntityScanner(applicationContext).scan(Document.class, Persistent.class));
-        context.setSimpleTypeHolder(aerospikeCustomConversions.getSimpleTypeHolder());
-        Class<?> fieldNamingStrategy = aerospikeDataProperties.getFieldNamingStrategy();
-        if (fieldNamingStrategy != null) {
-            context.setFieldNamingStrategy((FieldNamingStrategy) BeanUtils.instantiateClass(fieldNamingStrategy));
-        }
-        context.setDefaultNameSpace(aerospikeDataProperties.getNamespace());
-        return context;
-    }
-
-    @Bean(name = "aerospikeExceptionTranslator")
-    @ConditionalOnMissingBean(name = "aerospikeExceptionTranslator")
-    public AerospikeExceptionTranslator aerospikeExceptionTranslator() {
-        return new DefaultAerospikeExceptionTranslator();
-    }
-
 }
