+++ spring-boot-starter-data-aerospike-example/reactive/src/main/java/com/aerospike/example/reactive/Customer.java
@@ -1,0 +1,17 @@
+package com.aerospike.example.reactive;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import org.springframework.data.annotation.Id;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class Customer {
+    @Id
+    private String id;
+    private String firstName;
+    private String lastName;
+    private long age;
+}
