+++ spring-boot-starter-data-aerospike-example/reactive/src/main/java/com/aerospike/example/reactive/ReactiveCustomerRepository.java
@@ -1,0 +1,8 @@
+package com.aerospike.example.reactive;
+
+import org.springframework.data.aerospike.repository.ReactiveAerospikeRepository;
+import reactor.core.publisher.Flux;
+
+public interface ReactiveCustomerRepository extends ReactiveAerospikeRepository<Customer, String> {
+    Flux<Customer> findByLastNameOrderByFirstNameAsc(String lastName);
+}
