+++ spring-boot-starter-data-aerospike-example/reactive/src/main/java/com/aerospike/example/reactive/ReactiveSpringDataAerospikeExampleApplication.java
@@ -1,0 +1,13 @@
+package com.aerospike.example.reactive;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ReactiveSpringDataAerospikeExampleApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(ReactiveSpringDataAerospikeExampleApplication.class, args);
+	}
+
+}
