+++ spring-boot-starter-data-aerospike-example/reactive/src/main/resources/application.properties
