+++ spring-boot-starter-data-aerospike-example/reactive/src/test/resources/logback-test.xml
