--- spring-boot-starter-data-aerospike-example/src/main/java/org/springframework/boot/autoconfigure/data/aerospike/example/ReactiveSpringDataAerospikeExampleApplication.java
@@ -1,13 +1,0 @@
-package org.springframework.boot.autoconfigure.data.aerospike.example;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ReactiveSpringDataAerospikeExampleApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ReactiveSpringDataAerospikeExampleApplication.class, args);
-	}
-
-}
