--- spring-boot-starter-data-aerospike-example/src/main/java/org/springframework/boot/autoconfigure/data/aerospike/example/entity/Customer.java
@@ -1,17 +1,0 @@
-package org.springframework.boot.autoconfigure.data.aerospike.example.entity;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import org.springframework.data.annotation.Id;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class Customer {
-    @Id
-    private String id;
-    private String firstName;
-    private String lastName;
-    private long age;
-}
