--- spring-boot-starter-data-aerospike-example/src/main/java/org/springframework/boot/autoconfigure/data/aerospike/example/repository/CustomerRepository.java
@@ -1,10 +1,0 @@
-package org.springframework.boot.autoconfigure.data.aerospike.example.repository;
-
-import org.springframework.boot.autoconfigure.data.aerospike.example.entity.Customer;
-import org.springframework.data.aerospike.repository.AerospikeRepository;
-
-import java.util.List;
-
-public interface CustomerRepository extends AerospikeRepository<Customer, String> {
-    List<Customer> findByLastNameOrderByFirstNameAsc(String lastName);
-}
