--- spring-boot-starter-data-aerospike-example/src/test/java/org/springframework/boot/autoconfigure/data/aerospike/example/controller/SyncIntegrationTest.java
@@ -1,9 +1,0 @@
-package org.springframework.boot.autoconfigure.data.aerospike.example.controller;
-
-public class SyncIntegrationTest extends BaseIntegrationTest {
-
-    @Override
-    protected String urlPath() {
-        return "";
-    }
-}
