+++ spring-boot-starter-data-aerospike-example/sync/src/main/java/com/aerospike/example/sync/SyncCustomerRepository.java
@@ -1,0 +1,9 @@
+package com.aerospike.example.sync;
+
+import org.springframework.data.aerospike.repository.AerospikeRepository;
+
+import java.util.List;
+
+public interface SyncCustomerRepository extends AerospikeRepository<Customer, String> {
+    List<Customer> findByLastNameOrderByFirstNameAsc(String lastName);
+}
