+++ spring-boot-starter-data-aerospike-example/sync/src/main/java/com/aerospike/example/sync/SyncSpringDataAerospikeExampleApplication.java
@@ -1,0 +1,13 @@
+package com.aerospike.example.sync;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class SyncSpringDataAerospikeExampleApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(SyncSpringDataAerospikeExampleApplication.class, args);
+	}
+
+}
