+++ spring-boot-starter-data-aerospike-example/sync/src/main/resources/application.properties
@@ -1,0 +1,3 @@
+spring.aerospike.hosts=${embedded.aerospike.host}:${embedded.aerospike.port}
+spring.data.aerospike.namespace=${embedded.aerospike.namespace}
+spring.data.aerospike.scans-enabled=true
